Improve Cloudformation template to create endpoint
Add a previously generated training model as dependency to create Endpoint

Extract an abstract part of the integration tests. It will allow extend
possibilities to improve IT

Extract code which charge to generate sample data and send it to HTTP API

Current issues:

 * Policies in MLExecutionRole to permissive - need optimization

 * Hardcoded path to s3 bucket with ML model (it need for the
   endpoint) - need preprocess cloudformation template
