+++ fds-lambda-ml-integration/src/test/java/com/provectus/fds/ml/SageMakerApp.java
@@ -8,7 +8,7 @@ public class SageMakerApp {
     public static void main(String[] args) throws IOException {
         new JobRunner()
                 .createJob(
-                        new IntegrationModuleHelper(), true,
+                        new IntegrationModuleHelper(),
                         "s3://newfdsb/athena/train_4188522943897605003.csv",
                         "s3://newfdsb/athena/verify_8880021217306585575.csv");
     }
